refactor the method
