+++ evm-tests/test/staking.precompile.reward.test.ts
@@ -1,5 +1,5 @@
 import * as assert from "assert";
-import { getAliceSigner, getDevnetApi, getRandomSubstrateKeypair, waitForTransactionWithRetry } from "../src/substrate"
+import { getDevnetApi, getRandomSubstrateKeypair } from "../src/substrate"
 import { devnet } from "@polkadot-api/descriptors"
 import { TypedApi } from "polkadot-api";
 import { convertPublicKeyToSs58 } from "../src/address-utils"
@@ -8,7 +8,8 @@ import {
     forceSetBalanceToSs58Address, addNewSubnetwork, burnedRegister,
     setTxRateLimit, setTempo, setWeightsSetRateLimit, setSubnetOwnerCut, setMaxAllowedUids,
     setMinDelegateTake, setActivityCutoff, addStake, setWeight, rootRegister,
-    startCall
+    startCall,
+    disableAdminFreezeWindowAndOwnerHyperparamRateLimit
 } from "../src/subtensor"
 
 describe("Test neuron precompile reward", () => {
@@ -39,20 +40,7 @@ describe("Test neuron precompile reward", () => {
         await startCall(api, netuid, coldkey)
 
         console.log("test the case on subnet ", netuid)
-        // Disable admin freeze window and owner hyperparam rate limiting for tests
-        {
-            const alice = getAliceSigner()
-
-            // Set AdminFreezeWindow to 0
-            const setFreezeWindow = api.tx.AdminUtils.sudo_set_admin_freeze_window({ window: 0 })
-            const sudoFreezeTx = api.tx.Sudo.sudo({ call: setFreezeWindow.decodedCall })
-            await waitForTransactionWithRetry(api, sudoFreezeTx, alice)
-
-            // Set OwnerHyperparamRateLimit to 0
-            const setOwnerRateLimit = api.tx.AdminUtils.sudo_set_owner_hparam_rate_limit({ limit: BigInt(0) })
-            const sudoOwnerRateTx = api.tx.Sudo.sudo({ call: setOwnerRateLimit.decodedCall })
-            await waitForTransactionWithRetry(api, sudoOwnerRateTx, alice)
-        }
+        await disableAdminFreezeWindowAndOwnerHyperparamRateLimit(api)
 
         await setTxRateLimit(api, BigInt(0))
         await setTempo(api, root_netuid, root_tempo)
